move avs reader tests to reader.rs
